+++ src/locations/popup/EmailForm.cpp
@@ -102,7 +102,7 @@ void EmailForm::keyboardTouchSignalHandler()
 
 		currentEmail.clear();
 
-		for (int i = 0; i < codesVec.size(); i++)
+		for (size_t i = 0; i < codesVec.size(); i++)
 		{
 			char letter = codesVec[i][0];
 			letter = toupper(letter);
